requirements.txt
Flask==2.3.2
python-docx==0.8.11
nltk==3.8.1

